+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "xcore"
-version = "2.3.0"
+version = "2.3.1"
 description = "Plugin-first orchestration framework built on FastAPI"
 authors = [
     { name = "Eliezer Traore", email = "68350805+traoreera@users.noreply.github.com" },
@@ -13,7 +13,7 @@ dependencies = [
     # SDK (external)
     "xcoresdk @ git+https://github.com/xcore-team/xcoreSDK.git@main",
     # external cli
-    "xcoreCli @ git+https://github.com/xcore-team/xcoreCli.git@main",
+    "xcoreCli @ git+https://github.com/xcore-team/xcoreCli.git@main",  # main cassé (hatch wheel config manquante)
 
     # Web
     "fastapi[standard]>=0.135.1,<1.0.0",
