+++ xcore/__version__.py
@@ -1,4 +1,4 @@
-__version__ = "2.3.0"
-__version_info__ = (2, 3, 0)
+__version__ = "2.3.1"
+__version_info__ = (2, 3, 1)
 __author__ = "xcore contributors"
 __license__ = "MIT"
