+++ xcore/kernel/runtime/middlewares/__init__.py
@@ -1,9 +1,14 @@
-from .middleware import Middleware, MiddlewarePipeline
-from .middleware_registry import MiddlewareRegistry
-from .permissions import PermissionMiddleware
-from .ratelimit import RateLimitMiddleware
-from .retry import RetryMiddleware
-from .tracing import TracingMiddleware
+from xcore.kernel.middlewares import (
+    Middleware,
+    MiddlewarePipeline,
+    MiddlewareRegistry,
+    PermissionMiddleware,
+    RateLimitMiddleware,
+    RetryMiddleware,
+    TracingMiddleware,
+)
+
+from .ipc_auth import IPCAuthMiddleware
 
 __all__ = [
     "Middleware",
@@ -13,4 +18,5 @@ __all__ = [
     "RateLimitMiddleware",
     "RetryMiddleware",
     "TracingMiddleware",
+    "IPCAuthMiddleware",
 ]
