+++ xcore/kernel/sandbox/middlewares/__init__.py
@@ -1,9 +1,12 @@
-from .middleware import Middleware, MiddlewarePipeline
-from .middleware_registry import MiddlewareRegistry
-from .permissions import PermissionMiddleware
-from .ratelimit import RateLimitMiddleware
-from .retry import RetryMiddleware
-from .tracing import TracingMiddleware
+from xcore.kernel.middlewares import (
+    Middleware,
+    MiddlewarePipeline,
+    MiddlewareRegistry,
+    PermissionMiddleware,
+    RateLimitMiddleware,
+    RetryMiddleware,
+    TracingMiddleware,
+)
 
 __all__ = [
     "Middleware",
