+++ xcore/kernel/sandbox/middlewares/ratelimit.py
@@ -1,22 +1,1 @@
-"""
-ratelimit.py — Middleware for plugin rate limiting.
-"""
-
-from __future__ import annotations
-
-from ...sandbox.limits import RateLimitExceeded
-from .middleware import Middleware
-
-
-class RateLimitMiddleware(Middleware):
-    def __init__(self, rate):
-        self._rate = rate
-
-    async def __call__(
-        self, plugin_name, action, payload, next_call, handler, **kwargs
-    ):
-        try:
-            self._rate.check(plugin_name)
-        except RateLimitExceeded as e:
-            return {"status": "error", "msg": str(e), "code": "rate_limit_exceeded"}
-        return await next_call(plugin_name, action, payload, handler, **kwargs)
+from xcore.kernel.middlewares.ratelimit import *  # noqa: F401, F403
