+++ xcore/services/cache/backends/memory.py
@@ -94,9 +94,11 @@ class MemoryBackend:
     async def mset(self, mapping: dict[str, Any], ttl: int | None = None) -> None:
         """Définit plusieurs clés d'un coup."""
         effective_ttl = ttl if ttl is not None else self._ttl
-        expires_at = (time.monotonic() + effective_ttl) if effective_ttl > 0 else None
 
         for k, v in mapping.items():
+            expires_at = (
+                (time.monotonic() + effective_ttl) if effective_ttl > 0 else None
+            )
             if k in self._store:
                 self._store.move_to_end(k)
             self._store[k] = _Entry(value=v, expires_at=expires_at)
